Move modules that are used by both client and server into shared
