+++ nusight2/src/client/components/vision/network.ts
@@ -1,8 +1,8 @@
 import { action, runInAction } from 'mobx'
 import { UnreachableError } from '../../../shared/base/unreachable_error'
+import { fourcc, fourccToString } from '../../../shared/image_decoder/fourcc'
 import { message } from '../../../shared/messages'
 import { toSeconds } from '../../../shared/time/timestamp'
-import { fourcc, fourccToString } from '../../image_decoder/fourcc'
 import { Matrix4 } from '../../math/matrix4'
 import { Vector2 } from '../../math/vector2'
 import { Vector3 } from '../../math/vector3'
