+++ nusight2/src/client/image_decoder/image_decoder.ts
@@ -20,8 +20,8 @@ import { LinearFilter } from 'three'
 import { NearestFilter } from 'three'
 import { Camera } from 'three'
 
-import { fourccToString } from './fourcc'
-import { fourcc } from './fourcc'
+import { fourccToString } from '../../shared/image_decoder/fourcc'
+import { fourcc } from '../../shared/image_decoder/fourcc'
 import bayerFragmentShader from './shaders/bayer.frag'
 import bayerVertexShader from './shaders/bayer.vert'
