+++ nusight2/src/shared/base/tests/compose.tests.ts
