+++ nusight2/src/shared/image_decoder/fourcc.ts
