+++ nusight2/src/virtual_robots/simulators/vision_simulator.ts
@@ -2,8 +2,8 @@ import fs from 'fs'
 import { autorun } from 'mobx'
 import { computedFn } from 'mobx-utils'
 import { Matrix4, Vector3 } from 'three'
-import { fourcc } from '../../client/image_decoder/fourcc'
 import { FieldDimensions } from '../../shared/field/dimensions'
+import { fourcc } from '../../shared/image_decoder/fourcc'
 import { Imat4, message } from '../../shared/messages'
 import { NUClearNetClient } from '../../shared/nuclearnet/nuclearnet_client'
 import { toTimestamp } from '../../shared/time/timestamp'
